Fixed warnings in script templates
I was getting warnings in the editor with scripts derived from the
script templates because they didn't specify a return type for the
`_get_configuration_warnings()` function.

That said, this has created four warnings that I cannot understand
in VSCode's linting, however that works. So this is theoretically not
ready for incorporation without further review.

The errors I'm seeing are all:

    The function signature doesn't match the parent. Parent signature is "_get_configuration_warnings() -> Variant".

This makes no sense to me, because if I ctrl-click to the parent, it
clearly shows `-> PackedStringArray`. I'm seeing this for the following
files:

* `fsm_state_integrated_bt.gd`
* `fsm_state_integration_return.gd`
* `new_state.gd`
* `new_transition.gd`

The common link here is that these are FSM ones, and the BT ones didn't
give me any such issues, but I have no idea if that's relevant.

Also, I have the formatter installed, so various changes to whitespace
are a result of that.